Update README.md
The primary objective of our project is to build a predictive model which is capable of predicting house price in USA. The dataset used for this project, sourced from Kaggle, includes various features related to houses. The goal is to classify predict house price based on different features of a house like square feet, area, number of bedrooms, etc.